+++ app/src/main/java/de/thomaskuenneth/benice/AppChooserActivity.kt
@@ -180,14 +180,19 @@ class AppChooserActivity : ComponentActivity() {
         secondApp: AppInfo,
         delay: Long,
         label: String,
-        addDynamicShortcut: Boolean
+        addDynamicShortcut: Boolean,
+        layout: AppPairIconLayout
     ) {
         if (shortcutManager.isRequestPinShortcutSupported) {
             val id = "${firstApp.className}|${secondApp.className}"
             val shortcutInfo = ShortcutInfo.Builder(this, id).setIcon(
                 Icon.createWithBitmap(
                     createAppPairBitmap(
-                        firstApp = firstApp, secondApp = secondApp
+                        firstApp = firstApp,
+                        secondApp = secondApp,
+                        bigWidth = shortcutManager.iconMaxWidth,
+                        bigHeight = shortcutManager.iconMaxHeight,
+                        layout = layout
                     )
                 )
             ).setShortLabel(label).setIntent(createLaunchAppPairIntent(firstApp, secondApp, delay))
@@ -206,27 +211,33 @@ class AppChooserActivity : ComponentActivity() {
 }
 
 fun createAppPairBitmap(
-    firstApp: AppInfo, secondApp: AppInfo
+    firstApp: AppInfo, secondApp: AppInfo, bigWidth: Int, bigHeight: Int, layout: AppPairIconLayout
 ): Bitmap {
-    val bigWidth = 512
-    val bigHeight = 512
     val smallWidth = bigWidth / 2
     val smallHeight = bigHeight / 2
-    val y = smallHeight / 2F
+    val verticalMargin = smallHeight / 2F
+    val topFirstBitmap = if (layout == AppPairIconLayout.DIAGONAL) smallHeight.toFloat() else verticalMargin
+    val topSecondBitmap = if (layout == AppPairIconLayout.DIAGONAL) 0F else verticalMargin
     return Bitmap.createBitmap(bigWidth, bigHeight, Bitmap.Config.ARGB_8888).also { bitmap ->
         val bitmapPaint = Paint().also { paint ->
             paint.isAntiAlias = true
             paint.isFilterBitmap = true
         }
         Canvas(bitmap).run {
+            val firstBitmap = firstApp.icon.toBitmap(smallWidth, smallHeight)
+            val secondBitmap = secondApp.icon.toBitmap(smallWidth, smallHeight)
             drawPaint(Paint().also {
                 it.color = Color.TRANSPARENT
             })
-            drawBitmap(firstApp.icon.toBitmap(smallWidth, smallHeight), 0F, y, bitmapPaint)
+            save()
+            density = firstBitmap.density
+            drawBitmap(firstBitmap, 0F, topFirstBitmap, bitmapPaint)
+            restore()
+            density = secondBitmap.density
             drawBitmap(
-                secondApp.icon.toBitmap(smallWidth, smallHeight),
+                secondBitmap,
                 smallWidth.toFloat(),
-                y,
+                topSecondBitmap,
                 bitmapPaint
             )
         }