+++ app/src/main/java/de/thomaskuenneth/benice/AppPairIconLayout.kt
@@ -1,0 +1,5 @@
+package de.thomaskuenneth.benice
+
+enum class AppPairIconLayout {
+    HORIZONTAL, DIAGONAL
+}